+++ modi/common-security/build.gradle
@@ -8,3 +8,4 @@ jar {
     enabled = true
 }
 
+
