+++ modi/member-service/src/main/java/com/coc/modi/auth/presentation/MemberAuthController.java
@@ -4,7 +4,6 @@ import com.coc.modi.auth.application.MemberAuthService;
 import com.coc.modi.auth.application.EmailVerificationService;
 import com.coc.modi.auth.application.dto.MemberLoginResponse;
 import com.coc.modi.auth.presentation.dto.*;
-import com.coc.modi.common.ApiResponse;
 
 import lombok.RequiredArgsConstructor;
