+++ modi/member-service/src/main/java/com/coc/modi/member/domain/Member.java
@@ -1,5 +1,7 @@
 package com.coc.modi.member.domain;
 
+import com.coc.modi.account.member.domain.MemberRole;
+import com.coc.modi.account.member.domain.MemberStatus;
 import com.coc.modi.common.BaseEntity;
 
 import jakarta.persistence.*;
