+++ modi/member-service/src/main/java/com/coc/modi/member/presentation/MemberController.java
@@ -6,7 +6,6 @@ import com.coc.modi.member.application.dto.MemberSignupResponse;
 import com.coc.modi.member.presentation.dto.MemberSignupRequest;
 import com.coc.modi.member.presentation.dto.MemberUpdateRequest;
 import com.coc.modi.member.presentation.dto.MemberPasswordUpdateRequest;
-import com.coc.modi.common.ApiResponse;
 
 import lombok.RequiredArgsConstructor;
