+++ modi/modi-config/build.gradle
@@ -1,8 +1,26 @@
+plugins {
+	id 'java'
+	id 'org.springframework.boot' version '3.5.8'
+	id 'io.spring.dependency-management' version '1.1.7'
+}
+
+group = 'com.coc'
+version = '0.0.1-SNAPSHOT'
+description = 'modi-config'
+
+java {
+	toolchain {
+		languageVersion = JavaLanguageVersion.of(17)
+	}
+}
+
+repositories {
+	mavenCentral()
+}
 
 dependencies {
     // 공통 모듈 공유
-    implementation project(':common')
-
-    implementation 'org.springframework.cloud:spring-cloud-starter-gateway'
-
+    implementation 'org.springframework.boot:spring-boot-starter-actuator'
+    implementation 'org.springframework.cloud:spring-cloud-config-server'
+    implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
 }
