+++ modi/modi-config/src/main/java/com/coc/config/ModiConfigApplication.java
@@ -2,7 +2,9 @@ package com.coc.config;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.config.server.EnableConfigServer;
 
+@EnableConfigServer
 @SpringBootApplication
 public class ModiConfigApplication {
