+++ modi/modi-discovery/.gitattributes
@@ -1,0 +1,3 @@
+/gradlew text eol=lf
+*.bat text eol=crlf
+*.jar binary
