+++ modi/modi-discovery/src/main/java/com/coc/discovery/ModiDiscoveryApplication.java
@@ -2,7 +2,9 @@ package com.coc.discovery;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
 
+@EnableEurekaServer
 @SpringBootApplication
 public class ModiDiscoveryApplication {
