+++ modi/modi-gateway/build.gradle
@@ -1,8 +1,32 @@
+plugins {
+    id 'java'
+    id 'org.springframework.boot' version '3.5.8'
+    id 'io.spring.dependency-management' version '1.1.7'
+}
+
+group = 'com.coc'
+version = '0.0.1-SNAPSHOT'
+description = 'modi-gateway'
+
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(17)
+    }
+}
+
+repositories {
+    mavenCentral()
+}
 
 dependencies {
-    // 공통 모듈 공유
-    implementation project(':common')
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    implementation 'org.springframework.cloud:spring-cloud-starter-gateway-server-webflux'
+    implementation 'org.springframework.cloud:spring-cloud-starter-config'
+    implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
+    implementation 'org.springdoc:springdoc-openapi-starter-webflux-ui:2.8.8'
 
-    implementation 'org.springframework.cloud:spring-cloud-starter-gateway'
+}
 
+tasks.named('test') {
+    useJUnitPlatform()
 }
