+++ modi/modi-gateway/src/main/java/com/coc/gateway/config/OpenApiDocConfig.java
@@ -1,36 +1,36 @@
-package com.coc.gateway.config;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
-
-import org.springdoc.core.models.GroupedOpenApi;
-import org.springdoc.core.properties.SwaggerUiConfigParameters;
-import org.springframework.cloud.gateway.route.RouteDefinition;
-import org.springframework.cloud.gateway.route.RouteDefinitionLocator;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class OpenApiDocConfig {
-	
-	@Bean
-	public List<GroupedOpenApi> apis(SwaggerUiConfigParameters swaggerUiConfigParameters,
-									 RouteDefinitionLocator locator) {
-		
-		List<GroupedOpenApi> groups = new ArrayList<>();
-		List<RouteDefinition> definitions = locator.getRouteDefinitions().log("OpenApiDocConfig")
-				.collectList().block();
-		
-		Optional.ofNullable(definitions).stream().flatMap(Collection::stream)
-				.filter(routeDefinition -> routeDefinition.getId().matches(".*-service"))
-				.forEach(routeDefinition -> {
-					String name = routeDefinition.getId();
-					swaggerUiConfigParameters.addGroup(name, "/v3/api-docs/" + name);
-					GroupedOpenApi.builder().pathsToMatch("/" + name + "/**").group(name).build();
-				});
-		
-		return groups;
-	}
-}
+// package com.coc.gateway.config;
+//
+// import java.util.ArrayList;
+// import java.util.Collection;
+// import java.util.List;
+// import java.util.Optional;
+//
+// import org.springdoc.core.models.GroupedOpenApi;
+// import org.springdoc.core.properties.SwaggerUiConfigParameters;
+// import org.springframework.cloud.gateway.route.RouteDefinition;
+// import org.springframework.cloud.gateway.route.RouteDefinitionLocator;
+// import org.springframework.context.annotation.Bean;
+// import org.springframework.context.annotation.Configuration;
+//
+// @Configuration
+// public class OpenApiDocConfig {
+//
+// 	@Bean
+// 	public List<GroupedOpenApi> apis(SwaggerUiConfigParameters swaggerUiConfigParameters,
+// 									 RouteDefinitionLocator locator) {
+//
+// 		List<GroupedOpenApi> groups = new ArrayList<>();
+// 		List<RouteDefinition> definitions = locator.getRouteDefinitions().log("OpenApiDocConfig")
+// 				.collectList().block();
+//
+// 		Optional.ofNullable(definitions).stream().flatMap(Collection::stream)
+// 				.filter(routeDefinition -> routeDefinition.getId().matches(".*-service"))
+// 				.forEach(routeDefinition -> {
+// 					String name = routeDefinition.getId();
+// 					swaggerUiConfigParameters.addGroup(name, "/v3/api-docs/" + name);
+// 					GroupedOpenApi.builder().pathsToMatch("/" + name + "/**").group(name).build();
+// 				});
+//
+// 		return groups;
+// 	}
+// }
