+++ modi/product-service/src/main/java/com/coc/modi/product/exception/GlobalExceptionHandler.java
@@ -1,8 +1,6 @@
 package com.coc.modi.product.exception;
 
-import com.coc.modi.common.ApiResponse;
 import com.coc.modi.common.BaseException;
-import com.coc.modi.common.ErrorCode;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
