+++ modi/rental-service/src/main/java/com/coc/modi/rental/RentalServiceApplication.java
@@ -6,7 +6,7 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @EnableFeignClients
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = "com.coc.modi")
 public class RentalServiceApplication {
 
     public static void main(String[] args) {
