+++ modi/seller-service/src/main/java/com/coc/modi/seller/settlement/presentation/SettlementBatchInternalController.java
@@ -1,6 +1,5 @@
 package com.coc.modi.seller.settlement.presentation;
 
-import com.coc.modi.common.ApiResponse;
 import com.coc.modi.seller.settlement.application.SettlementBatchService;
 import com.coc.modi.seller.settlement.application.dto.SettlementBatchCreateCommand;
 import com.coc.modi.seller.settlement.application.dto.SettlementBatchResponse;
