+++ modi/settings.gradle
@@ -2,6 +2,7 @@ rootProject.name = 'modi'
 
 include(
         'common',
+        'common-security',
         'account-service',
         'product-service',
         'rental-service',
